Remove redundant stuff

+++ scripts/99-cleanup.sh
@@ -6,6 +6,6 @@ shopt -s extglob
 dnf5 clean all
 
 rm -rf /tmp/* || true
-rm -rf /var/!(cache)
+rm -rf /"${var:?}"/!(cache)
 rm -rf /var/cache/!(rpm-ostree)
 rm -rf /scripts
